Changed allocated budget type to decimal

+++ src/DAL/OrganizationModel/Entities/Segment.cs
@@ -16,7 +16,7 @@ namespace Tayra.Models.Organizations
         [MaxLength(2000)]
         public string Avatar { get; set; }
 
-        public int AllocatedBudget { get; set; }
+        public decimal AllocatedBudget { get; set; }
 
         [MaxLength(50)]
         public string Timezone { get; set; }
